Toggle main UI whether recording is enabled or not
The notice saying that recording is disabled ("disable notice") is placed on
top of the buttons and set to hidden by default. When the audio session is
initialized and recording is enabled, it would show the buttons. When recording
is not allowed, it would unhide the notice and hide the buttons.

The cool thing: the app automatically listens to the privacy settings and the
UI automatically changes depending on the status.