Overhaul recording and playback internals
Realized that wiring up UISlider with AVQueuePlayer is going to be
a challenge (many AVPlayerItems, each with their own length info
and position in the list that needs to be coordinated).

Instead,
+ made local composition in `concatChunks()` global along with
+ `insertAt` (CMTimeRange) and
+ changed `articleChunks` ([AVURLAsset]) to `latestChunk` (AVURLAsset)
+ and divided `concatChunks()` into `appendChunk()` and `exportArticle()`

`appendChunk()`
 1) adds the `latestChunk` (i.e., the recording since the
last stop/pause up to the now) to `articleSoFar` (AVMutableComposition
that contains all the article chunks)
2) keeps track of previous chunks (AVURLAssets) in `leftoverChunks`
3) `nil`s out `latestChunk`

`zeroAudioArtifacts()`
makes sure that all global audio objects are set back to their
initial positions (used on startup, in "Start Over" button and
when exporting audio).