Set up Pause button during Playback
Fundamental issues in the app:
==============================

The timer label is set up to change on every centisecond via a Timer that
starts firing on recording and playback. It's state is also saved when
switching to playback, and restored when playback is stopped. During playback
another timer is used (AVPlayer's `addPeriodicTimeObserver`), and the problem
is that the end result of the 2 timers are not the same. For example,
recording timer shows "00:00:07" and at different times, playback timer would
show "00:00:06" or "00:00:08", seeemingly at random.

The two timers should be in sync, and I assume that the way to do it is
getting time data from the recording itself or player item during playback.