Mitigating #14 via GCD
Adding a serial dispatch queue to synchronize the two async operations,
`zeroRecordArtifacts` and uploading the result to Firebase. Not a full
solution though because if subsequent recordings started fast enough,
the previous counters are still in play and would be invalidated during
user actions (e.g., recording), resulting in crash.